entity_extraction/requirements.txt
absl-py==0.11.0
arrow==0.17.0
astunparse==1.6.3
bert-tensorflow==1.0.4
binaryornot==0.4.4
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cycler==0.10.0
et-xmlfile==1.0.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
jdcal==1.4.1
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
logging==0.4.9.6
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
nltk==3.5
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.4.0.46
openpyxl==3.0.5
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.0.1
poyo==0.5.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn==0.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.10.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
tokenization==1.0.7
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
zipp==3.4.0
